requirements.txt
python-dotenv
openai>=1.98.0
feedparser
python-slugify==8.*
Unidecode==1.3.*
requests>=2.31
